Stop loading so much of the plugin before we know we're ready
Plugin could be half-activated, due to secret's absence, but Events Store would still capture data. As the rest of the plugin, notably its CLI commands, aren't loaded, this could store data into an unreachable hole. Also, with the introduction of a custom table, the table shouldn't be created when the environment isn't ready.