Remove remnant of CPT that imposes problematic and arbitrary limit
When events are capped in this way, if they don't appear in the reconstructed array, they can be continually rescheduled.

With the introduction of the block on existing-event creation introduced in 61597bc, this approach should longer be necessary. The creation of duplicates is also greatly reduced by the table's unique key.

Fixes #94