Reduce number of queries from get_option
By reducing the number of queries by 50x, we will reduce the time spent in this function and increase performance. I'm going with 5000 because we consider it an upper limit for the number of jobs you can have. If sites have more than 5000 pending events, cron gets disabled and there are some alerts.