Add command to list sites divided up by host
Previously the runner would use `wp site list` to get a list of sites
to check. The problem is that if you have many runners (i.e. many cron
hosts per site) we waste a lot of time trying to run the same jobs
on every host.

The idea here is to dedicate certain sites to certain hosts.

The runner will run `wp cron-control orchestrate sites list` to get a
list of sites. The act of running that command triggers a "heartbeat",
which stores a list of all hosts in the object cache along with the
last heartbeat time.

We use the list of hosts to divide sites into groups. Right now the idea
is that each site will have 2 hosts dedicated to running cron for it
so that if a host goes away, we have a bit of time to recover.

If a host is not seen for 3 heartbeats, it is removed from the list.