Fix errors for `wp orchestrate site list`
When building the site list for the orchestrate command, don't use home or siteurl since those don't always match up with the `wp_blogs` entry. That can result in a "site not found" when passed via the `--url` WP-CLI param and break cron.

Instead, construct the URL from data in the `wp_blogs` table.