Site List: Remove unnecessary switch_to_blog
The `get_home_url()` function calls `switch_to_blog()` and `restore_current_blog()` internally.

First, `restore_current_blog()` is basically `switch_to_blog()` in reverse. In this context, we don't actually care about switching back. We only want this one option for every site, so switching back is a waste of time. After we get the option, we're just outputting it.

More importantly though, `switch_to_blog()` is expensive and not necessary. We're only using the home URL for the scheme, which is unnecessary for a "URL" that gets passed to WP-CLI. The domain + path is enough (i.e. `wp --url=example.com/subsite`). It also seems a little silly, in this context at least, to do a full `switch_to_blog()` to query this one database option. It seems like we just need to know the table name and from there could query the options table directly. The expensive part of `switch_to_blog()` is where we re-initialize the object cache, thus wiping out all auto-loaded options.