Bypass local cache before writing to cache
The job cache uses one cache key to make the best use of the local PHP
cache. One request to memcached gets all the cached jobs and then we
can just request from the local cache for future requests.

This presents an issue where updates to the cache can write old data
back to memcached.

In order to avoid writing old data, we need to bypass the local cache
and read directly from memcached in our set and delete wrappers.