Fix scheduling future jobs
We see a lot of future jobs trying to and failing to run. In PHP, we were listing jobs that are due in the next 60 seconds, but in the Go runner we refuse to run events that are scheduled in the future.

https://github.com/Automattic/Cron-Control/blob/9928da3a5f9a78e6616cb82663c03f32ceef5845/runner/runner.go#L417

This is consuming a lot of workers with running jobs that are due in the next 60 seconds in some cases.

This PR synchronizes that offset to +0 seconds in the future.