Preserve keys when slicing the array (#166)
In PHP, the array union operator (+) ignores even numerical keys from the second array if those are already present in the left hand one. Thus in case we are slicing an array w/o preserving keys, we are then trying to merge two arrays with identical numerical keys. And thus, no union is happening.

This commit is fixing the issue by preserving the keys. Another approach would be to use `array_merge` which is appending identical numeric keys instead of ignoring those, but preserving the keys feels like a better approach, when we are dealing with cached and sliced array.

This should make #165 obsolete